OSD: respond to updates to ModuleSettings
This is useful because when running as a standalone OSD the
modules are enabled when messages arrived to make things
"just work".

In addition, getting the sttings at the beginning of Initialize
is very unsafe. This depends on the order modules are initialized
that the object even has been initialized. Never fetch objects
in the Initialize methods.