OSD: More detailed alarms.
This is based on the work I did in MSP in TL proper.  It's a slightly
more condensed form to store the names of alarms, but should fail to
compile if new alarms are added and not described here, provides more
detail on system configuration and manual control alarms, and considers
the boot cause to be an alarm for a bit at startup so you can see why it
booted.

Signed-off-by: James Cotton <peabody124@gmail.com>